refactor codes to incorporate both new and old interface
