+++ packages/openfermion/quri_parts/openfermion/utils/post_selection_filters.py
@@ -51,7 +51,7 @@ def create_bk_electron_number_post_selection_filter_fn(
     by Bravyi-Kitaev transformation.
     """
     inv_st_mapper = bravyi_kitaev.get_inv_state_mapper(
-        bravyi_kitaev.n_spin_orbitals_required(qubit_count)
+        bravyi_kitaev.n_spin_orbitals(qubit_count)
     )
 
     def filter_fn(bits: int) -> bool:
@@ -74,7 +74,7 @@ def create_scbk_electron_number_post_selection_filter_fn(
     by symmetry-conserving Bravyi-Kitaev transformation.
     """
     inv_st_mapper = symmetry_conserving_bravyi_kitaev.get_inv_state_mapper(
-        symmetry_conserving_bravyi_kitaev.n_spin_orbitals_required(qubit_count),
+        symmetry_conserving_bravyi_kitaev.n_spin_orbitals(qubit_count),
         n_electrons,
         sz,
     )
